pyproject.toml
[project]
name = "war-map-strategy"
version = "0.1.0"
description = "둠스데이 라스트서바이버 전쟁맵 전략 웹사이트"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.40.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.0.0",
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
